refactor(RecursiveAutomaton): move recursive automaton and related classes to distinct package
